// React/app/GerAdv_TS/ProcessOutputEngine/Interfaces/interface.ProcessOutputEngine.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IProcessOutputEngine {
// 202501251
    id: number;
 
	nome: string,
	database: string,
	tabela: string,
	campo: string,
	valor: string,
	output: string,
	administrador: boolean,
	outputsource: number,
	disableditem: boolean,
	idmodulo: number,
	isonlyprocesso: boolean,
	myid: number,
}

export interface IProcessOutputEngineFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IProcessOutputEngineIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}